refactor: google sign in and next auth directory structure (#178)
