+++ src/providers/googleSignInAuthentication/types.ts
@@ -1,6 +1,6 @@
 import { ElementType, ReactNode } from "react";
 
-export interface GoogleSignInProviderProps {
+export interface GoogleSignInAuthenticationProviderProps {
   APIServicesProvider: ElementType;
   children: ReactNode | ReactNode[];
   timeout?: number;
