+++ src/providers/nextAuthAuthentication/hooks/useNextAuthService.ts
@@ -1,8 +1,8 @@
 import { useCallback } from "react";
-import { transformRoute } from "../../../../hooks/authentication/session/useSessionActive";
-import { useRouteHistory } from "../../../../hooks/useRouteHistory";
-import { Service } from "../../auth/types";
-import { ProviderId } from "../../common/types";
+import { transformRoute } from "../../../hooks/authentication/session/useSessionActive";
+import { useRouteHistory } from "../../../hooks/useRouteHistory";
+import { Service } from "../../authentication/auth/types";
+import { ProviderId } from "../../authentication/common/types";
 import { service } from "../service/service";
 
 export const useNextAuthService = (): Service => {
