+++ src/providers/nextAuthAuthentication/types.ts
@@ -1,7 +1,7 @@
 import { Session } from "next-auth";
 import { ReactNode } from "react";
 
-export interface NextAuthProviderProps {
+export interface NextAuthAuthenticationProviderProps {
   children: ReactNode | ReactNode[];
   session?: Session | null;
   timeout?: number;
